+++ nav/approach.go
@@ -391,7 +391,7 @@ func (nav *Nav) InterceptApproach(airport string, lg *log.Logger) av.CommandInte
 	}
 }
 
-func (nav *Nav) AtFixCleared(fix, id string) av.CommandIntent {
+func (nav *Nav) AtFixCleared(fix, id string, straightIn bool) av.CommandIntent {
 	if nav.Approach.AssignedId == "" {
 		return av.MakeUnableIntent("unable. you never told us to expect an approach")
 	}
@@ -419,11 +419,15 @@ func (nav *Nav) AtFixCleared(fix, id string) av.CommandIntent {
 	if nav.Approach.AtFixClearedRoute == nil {
 		return av.MakeUnableIntent("unable. {fix} is not on the {appr} approach", fix, ap.FullName)
 	}
+	if straightIn && len(nav.Approach.AtFixClearedRoute) > 0 {
+		nav.Approach.AtFixClearedRoute[0].SetNoPT(true)
+	}
 
 	return av.ApproachIntent{
 		Type:         av.ApproachAtFixCleared,
 		ApproachName: ap.FullName,
 		Fix:          fix,
+		StraightIn:   straightIn,
 	}
 }