+++ whatsnew.md
@@ -3,6 +3,7 @@
   - Handle "at or above/at or below" speed restrictions more reasonably (don't treat them all as "at")
   - Added support for intercepting RNAV t-segments
   - Added support for "join the approach" for non-ILS/LOC approaches
+  - Fixed a few bugs with "at (fix), cleared straight in (approach)"
 - Added multiple new control instructions:
   - "after FIX, climb/descend and maintain ALT"
   - "after FIX, reduce/maintain/increase SPEED"